# fix: faster stream verification