@H4ad 2023-04-04
I was reading the code, check if I'm not wrong, this piece of code:

https://github.com/npm/ssri/blob/8e80eca497bd7f6a714498df6ad64f3faef3bc5e/lib/index.js#L82-L86

And then, the implementation of `match` is:

https://github.com/npm/ssri/blob/8e80eca497bd7f6a714498df6ad64f3faef3bc5e/lib/index.js#L283-L298

What this code is doing:

- Hashing the stream with all the algorithms (from opts or default 512).
- Then, getting the digest and parsing it again to transform to `Integrity` or `Hash`.
- Then, picking just one of the algorithms (basically wasting the time spent to parse it because we just need one).
- And finally, we compare the digest (that we already had)

If the assumptions are:

- We need to compare all the hashes and make sure all match.

So, the code is not implemented correctly.

If the assumptions are:

- We need to compare at least one hash.

So, the code is doing more than needed, we could just parse the sent digest, pick one, and then parse the stream for just that hash.

In the final, I think we could remove entirely the function of `pickAlgorithm` because the assumption behind it is wrong.