@wraithgar 2023-04-04
`pickAlgorithm` is a very very useful function still. Our package fetching library and caching library use it to pull out the "best" algorithm to use when reading fetching/reading data.

Reusing it in `match` I think was where we went wrong.

As far as what `match` *should* be doing, I think I need to think through a boolean chart.

Integrity can have one hash or multiple hashes,  sri can have one hash or multiple hashes.

`Integrity.match(sri)`: 

`(one hash).match(one hash)` - Hash algorithms should match, and digests should match.
`(one hash).match(multiple hashes)` - Integrity hash algorithm should be found in sri hashes, and the digests of that hash should match.
`(multiple hashes).match(one hash)` - sri hash algorithm should be found in Integrity hashes, and the digests of that hash should match.
`(multiple hashes).match(multiple hashes)` - Either we find the "best" algorithm of any matching hashes, and the digests of that one hash should match, or we find all matching algorithms between the Integrity and sri, and all of the digests should match.