@H4ad 2023-04-04
@wraithgar Okay, I'll look into this more over the weekend to see if I can get more optimizations out of this.

But those optimizations will be along with the `match` fix, so I guess we could separate it into another PR.

About this PR, the only thing missing is the tests, right? For the bug on `match` we will just forward it to another PR and keep it as far as I can understand.