@H4ad 2023-04-06
@jakebailey The PR that changes the behavior of `toString` is this one: https://github.com/npm/ssri/pull/75

Essentially, if the hash was created from a string like `sha256-xxx sha512-xxx`, the output when calling `parsed.toString()` is `sha512-xxx sha-256xxx` only when `strict=true`.

The only time `PNPM` call that method is in this place: https://github.com/search?q=repo%3Apnpm%2Fpnpm%20ssri.parse&type=code

So I think it will not be a problem for PNPM, for other package managers, I don't know if could be a problem because usually the integrity of the packages is just stored using a single algorithm, this little breaking change only affects when we see more than one algorithm with `strict=true`.