@wraithgar 2023-04-06
Yes that's correct. As long as folks are using the literal stringified representation but are instead letting ssri parse it they will be fine.